# [3/3] Optimize Slime Update Weights: Load Weight in non-blocking mode